refactored model
