+++ digitalpy/core/persistence/impl/default_persistence_facade.py
@@ -9,12 +9,12 @@
 #######################################################
 from typing import Any
 from digitalpy.core.event_manager import EventManager
-from digitalpy.model.object_id import ObjectId
-from digitalpy.model.persistence_facade import PersistenceFacade
-from digitalpy.model.persistent_object import PersistentObject
-from digitalpy.model.build_depth import BuildDepth
-from digitalpy.model.paging_info import PagingInfo
-from digitalpy.model.persistence_mapper import PersistenceMapper
+from digitalpy.core.persistence.object_id import ObjectId
+from digitalpy.core.persistence.persistence_facade import PersistenceFacade
+from digitalpy.core.persistence.persistent_object import PersistentObject
+from digitalpy.core.persistence.build_depth import BuildDepth
+from digitalpy.core.persistence.paging_info import PagingInfo
+from digitalpy.core.persistence.persistence_mapper import PersistenceMapper
 from digitalpy.core.state_change_event import StateChangeEvent
