+++ digitalpy/core/persistence/node_serializer.py
