+++ digitalpy/core/persistence/persistence_facade.py
@@ -8,11 +8,11 @@
 # 
 #######################################################
 from typing import Any
-from digitalpy.model.object_id import ObjectId
-from digitalpy.model.persistence_mapper import PersistenceMapper
+from digitalpy.core.persistence.object_id import ObjectId
+from digitalpy.core.persistence.persistence_mapper import PersistenceMapper
 from abc import ABC, abstractmethod
-from digitalpy.model.build_depth import BuildDepth
-from digitalpy.model.paging_info import PagingInfo
+from digitalpy.core.persistence.build_depth import BuildDepth
+from digitalpy.core.persistence.paging_info import PagingInfo
