+++ digitalpy/core/persistence/persistent_object_proxy.py
@@ -1,4 +1,4 @@
-from digitalpy.model.persistent_object import PersistentObject
+from digitalpy.core.persistence.persistent_object import PersistentObject
 
 
 class PersistentObjectProxy(PersistentObject):
