+++ digitalpy/core/persistence/value_change_event.py
@@ -1,5 +1,5 @@
 from digitalpy.core.event import Event
-from digitalpy.model.persistent_object import PersistentObject
+from digitalpy.core.persistence.persistent_object import PersistentObject
 
 class ValueChangeEvent(Event):
     NAME = "ValueChangeEvent"
